Update on pages and tabs
