+++ app/(home)/index.tsx
@@ -1,5 +1,6 @@
 
-import { StyleSheet, Text, View } from 'react-native';
+import { StyleSheet, Text, View} from 'react-native';
+
 import styles from "../../components/styles"
 import Button from "../../components/Button"
 import Input from "../../components/Input"
@@ -7,9 +8,9 @@ import Template from "../../components/Template"
 
 const Home = () => {
   return (
-    <View style={styles.container}>
-      <Text style={styles.welcometitle}>Calgary</Text>
-    </View>
+    <Template>
+      <Text style={styles.welcometitle}>Welcome</Text>
+    </Template>
   );
 }
