--- app/home/index.tsx
@@ -1,16 +1,0 @@
-
-import { Text, View } from 'react-native';
-import styles from "../../components/styles";
-
-
-const Home = () => {
-  return (
-    <View style={styles.container}>
-      <Text style={styles.welcometitle}>Welcome to Explore Canada</Text>
-    </View>
-  );
-}
-
-export default Home;
-
-
