+++ components/Template.tsx
@@ -1,15 +1,18 @@
-import {View, ScrollView, SafeAreaView} from "react-native"
+import {View, ScrollView, StatusBar} from "react-native"
+import {SafeAreaView} from 'react-native-safe-area-context'
+import styles from "./styles"
 type TemplateProps = {
     children:any;
 }
 
 const Template = ({children, ...props}:TemplateProps) => {
     return(
-        <SafeAreaView>
-            <ScrollView>
+      
+        <>
+            <ScrollView style={styles.scrView}>
                 <View>{children}</View>
             </ScrollView>
-        </SafeAreaView>
+          </>
     )
 }
